petal/backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1,<0.115.0
uvicorn[standard]>=0.24.0,<0.31.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.7.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# External APIs
anthropic>=0.8.1,<1.0.0
mem0ai>=0.0.10
supabase>=2.3.0,<3.0.0
openai>=1.0.0,<2.0.0

# Utilities
httpx>=0.25.2
python-multipart>=0.0.6

# Caching
redis>=7.0.0,<8.0.0
